requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
cloudinary==1.44.0
cryptography==44.0.2
Django==5.2
django-ckeditor==6.7.2
django-cors-headers==4.7.0
django-js-asset==3.1.2
django-oauth-toolkit==3.0.1
djangorestframework==3.16.0
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
jwcrypto==1.5.6
oauthlib==3.2.2
packaging==25.0
pillow==11.2.1
pycparser==2.22
PyMySQL==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
